requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
tqdm>=4.66.0
numpy>=1.26.4
pandas>=2.2.2
sentence-transformers>=2.7.0
chromadb>=0.5.3
pypdf>=4.2.0
httpx>=0.27.0
scikit-learn>=1.4.2
rank-bm25>=0.2.2

